修复代码审查发现的 10 个问题
- CORS: allow_origins 改为 localhost:3000，避免与 credentials 冲突
- 登录限速: 同一 IP 5 次失败锁定 60 秒
- 充值竞态: 使用原子 WHERE status=pending 防止双重加余额
- Float→Numeric: 金额字段改用 Numeric(10,2) 避免舍入误差
- billing: days 参数修复，传入 SQL 查询
- CSS: 补充 --bg-tertiary 变量和 pulse 动画
- register: 补充 credentials:include
- API Key: 脱敏显示，仅创建时返回完整密钥
- admin 页面: 移除 mock 数据回退，精简冗余代码

+++ backend/app/models/user.py
@@ -2,7 +2,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import String, Boolean, Float, DateTime, Text
+from sqlalchemy import String, Boolean, Float, Numeric, DateTime, Text
 from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.dialects.sqlite import TEXT
 
@@ -16,7 +16,7 @@ class User(Base):
     email: Mapped[str] = mapped_column(String(255), unique=True, index=True)
     username: Mapped[str] = mapped_column(String(100), unique=True, index=True)
     hashed_password: Mapped[str] = mapped_column(String(255))
-    balance: Mapped[float] = mapped_column(Float, default=0.0)
+    balance: Mapped[float] = mapped_column(Numeric(10, 2), default=0.0)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     is_admin: Mapped[bool] = mapped_column(Boolean, default=False)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
@@ -45,8 +45,8 @@ class ModelConfig(Base):
     upstream_base_url: Mapped[str] = mapped_column(String(500))  # 上游 API 地址
     upstream_api_key: Mapped[str] = mapped_column(String(500))  # 上游 API Key
     model_name: Mapped[str] = mapped_column(String(200))  # 上游模型名
-    price_per_input: Mapped[float] = mapped_column(Float, default=0.0)  # 每1M输入token价格
-    price_per_output: Mapped[float] = mapped_column(Float, default=0.0)  # 每1M输出token价格
+    price_per_input: Mapped[float] = mapped_column(Numeric(10, 2), default=0.0)  # 每1M输入token价格
+    price_per_output: Mapped[float] = mapped_column(Numeric(10, 2), default=0.0)  # 每1M输出token价格
     model_type: Mapped[str] = mapped_column(String(50), default="chat")  # chat/image/embedding/rerank/audio/video
     scene_tags: Mapped[str] = mapped_column(String(500), default="")  # 逗号分隔场景标签
     context_length: Mapped[int] = mapped_column(default=0)  # 上下文窗口
@@ -69,7 +69,7 @@ class UsageLog(Base):
     model_name: Mapped[str] = mapped_column(String(100))
     prompt_tokens: Mapped[int] = mapped_column(default=0)
     completion_tokens: Mapped[int] = mapped_column(default=0)
-    cost: Mapped[float] = mapped_column(Float, default=0.0)
+    cost: Mapped[float] = mapped_column(Numeric(10, 4), default=0.0)
     ip_address: Mapped[str] = mapped_column(String(45), nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
 
@@ -81,7 +81,7 @@ class RechargeOrder(Base):
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
     user_id: Mapped[int] = mapped_column(index=True)
     order_no: Mapped[str] = mapped_column(String(64), unique=True, index=True)
-    amount: Mapped[float] = mapped_column(Float)
+    amount: Mapped[float] = mapped_column(Numeric(10, 2))
     pay_method: Mapped[str] = mapped_column(String(20))  # alipay, wechat
     status: Mapped[str] = mapped_column(String(20), default="pending")  # pending, success, failed
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
